+++ src/components/Lesson/LessonOverview/LessonOverview.page.tsx
@@ -193,6 +193,9 @@ export function LessonOverview({ lesson }: LessonOverviewProps) {
                       signLanguageVideo={videoWithSignLanguageMuxPlaybackId}
                       title={lessonTitle}
                       transcriptSentences={transcriptSentences}
+                      temporaryUsePublicVideos={
+                        !isSlugLegacy(programmeSlug ?? "")
+                      }
                     />
                   </LessonItemContainer>
                 )}