+++ src/components/UnitModal/UnitModal.test.tsx
@@ -1,0 +1,59 @@
+import UnitModal from "./UnitModal";
+import { mockUnit, mockOptionalityUnit } from "./UnitModal.fixture";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
+describe("Unit modal", () => {
+  test("renders with correct heading", () => {
+    const { getByText } = renderWithTheme(<UnitModal unitData={mockUnit} />);
+    expect(getByText("Composition of numbers 6 to 10")).toBeInTheDocument();
+  });
+
+  test("renders the correct number of threads", () => {
+    const { getAllByTestId, getByText } = renderWithTheme(
+      <UnitModal unitData={mockUnit} />,
+    );
+    const testThread = getByText("Number: Addition and Subtraction");
+    const testThread2 = getByText("Number");
+
+    expect(getAllByTestId("thread-tag")).toHaveLength(2);
+    expect(testThread).toBeInTheDocument();
+    expect(testThread2).toBeInTheDocument();
+  });
+
+  test("lesson metadata renders correct data", () => {
+    const { getByText } = renderWithTheme(<UnitModal unitData={mockUnit} />);
+
+    expect(getByText("Maths")).toBeInTheDocument();
+    expect(getByText("Year 1")).toBeInTheDocument();
+  });
+
+  describe("non-optional units", () => {
+    test("does not render optionality card", () => {
+      const { queryAllByTestId } = renderWithTheme(
+        <UnitModal unitData={mockUnit} />,
+      );
+
+      expect(queryAllByTestId("unit-option-card")).toHaveLength(0);
+    });
+  });
+
+  describe("optional units", () => {
+    test("optionality cards render", () => {
+      const { getByTestId } = renderWithTheme(
+        <UnitModal unitData={mockOptionalityUnit} />,
+      );
+
+      const optionalityCard = getByTestId("unit-options-card");
+      expect(optionalityCard).toBeInTheDocument();
+    });
+
+    test("optionality cards render correct number of units", () => {
+      const { getAllByTestId } = renderWithTheme(
+        <UnitModal unitData={mockOptionalityUnit} />,
+      );
+
+      expect(getAllByTestId("unit-option")).toHaveLength(3);
+    });
+  });
+});