+++ src/node-lib/curriculum-api-2023/queries/lessonListing/lessonListing.schema.ts
@@ -9,8 +9,10 @@ const lessonListingSchema = z.object({
   unitSlug: z.string(),
   unitTitle: z.string(),
   tierSlug: z.string().nullish(),
+  tierTitle: z.string().nullish(),
   examBoardSlug: z.string().nullish(),
   examBoardTitle: z.string().nullish(),
+  yearTitle: z.string().nullish(),
   lessons: z.array(
     z.object({
       lessonSlug: z.string(),